test git push

+++ exercises/variables/variables3.rs
@@ -5,6 +5,6 @@
 
 
 fn main() {
-    let x: i32 = 10;
+    let x: i128 = 10;
     println!("Number {}", x);
 }
